fix(notifications): match db_call_id to BIGINT in Postgres outbox schema
Postgres calls.id is BIGSERIAL (bigint); INTEGER FK created a type
mismatch inconsistent with notification_send_log.call_id BIGINT.

+++ database/migrations/2026-05-12-notification-outbox.pgsql.sql
@@ -6,7 +6,7 @@
 -- consumed by FileWatcher's outbox tick.
 CREATE TABLE IF NOT EXISTS notification_outbox (
     id                  BIGSERIAL PRIMARY KEY,
-    db_call_id          INTEGER NOT NULL REFERENCES calls(id) ON DELETE CASCADE,
+    db_call_id          BIGINT NOT NULL REFERENCES calls(id) ON DELETE CASCADE,
     channel_id          BIGINT NOT NULL REFERENCES notification_channels(id) ON DELETE CASCADE,
     intent              VARCHAR(16) NOT NULL,
     resend_all          BOOLEAN NOT NULL DEFAULT FALSE,
